requirements/dev.txt
-r base.txt
psycopg2-binary==2.9.9
django-browser-reload==1.12.1
coreapi==2.3.3